requirements.txt
fastapi==0.104.1
uvicorn==0.24.0.post1
python-multipart==0.0.6
pydantic==2.4.2
pydantic[email]
requests==2.31.0
numpy==1.26.1
PyPDF2==3.0.1
python-dotenv
landingai-ade
beautifulsoup4==4.12.2
supabase
email-validator